pyproject.toml
[project]
name = "fisseq-qc-filter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { email = "gwenneth.straub@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3",
    "polars>=1.39.3",
]

[project.scripts]
fisseq-qc-filter = "fisseq_qc_filter:main"

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "isort>=8.0.1",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
